Model/model-recommend/requirements.txt
fastapi[all]
uvicorn
pandas
numpy
scikit-learn
requests
nltk
python-dotenv
